pyproject.toml
[project]
name = "FastAPITemplate"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    { name = "Korionov Dmitriy", email = "koriond98@gmail.com" },
]
dependencies = [
    "alembic==1.13.2",
    "annotated-types==0.7.0",
    "anyio==4.4.0",
    "async-timeout==4.0.3",
    "asyncpg==0.29.0",
    "bcrypt==4.0.1",
    "certifi==2024.7.4",
    "cffi==1.16.0",
    "cfgv==3.4.0",
    "click==8.1.7",
    "cryptography==42.0.8",
    "distlib==0.3.8",
    "dnspython==2.6.1",
    "ecdsa==0.19.0",
    "email-validator==2.2.0",
    "factory-boy==3.3.0",
    "Faker==26.0.0",
    "fastapi==0.111.1",
    "fastapi-cli==0.0.4",
    "fastapi-permissions==0.2.7",
    "filelock==3.15.4",
    "greenlet==3.0.3",
    "h11==0.14.0",
    "httpcore==1.0.5",
    "httptools==0.6.1",
    "httpx==0.27.0",
    "identify==2.6.0",
    "idna==3.7",
    "iniconfig==2.0.0",
    "Jinja2==3.1.4",
    "Mako==1.3.5",
    "markdown-it-py==3.0.0",
    "MarkupSafe==2.1.5",
    "mdurl==0.1.2",
    "msgspec==0.18.6",
    "nodeenv==1.9.1",
    "packaging==24.1",
    "platformdirs==4.2.2",
    "pluggy==1.5.0",
    "pre-commit==3.7.1",
    "pyasn1==0.6.0",
    "pycparser==2.22",
    "pydantic==2.8.2",
    "pydantic-settings==2.3.4",
    "pydantic-core==2.20.1",
    "Pygments==2.18.0",
    "pytest==8.2.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-jose==3.3.0",
    "python-multipart==0.0.9",
    "PyYAML==6.0.1",
    "rich==13.7.1",
    "rsa==4.9",
    "ruff==0.5.2",
    "shellingham==1.5.4",
    "six==1.16.0",
    "sniffio==1.3.1",
    "SQLAlchemy==2.0.31",
    "starlette==0.37.2",
    "typer==0.12.3",
    "typing-extensions==4.12.2",
    "ujson==5.10.0",
    "uvicorn==0.30.1",
    "uvloop==0.19.0",
    "virtualenv==20.26.3",
    "watchfiles==0.22.0",
    "websockets==12.0",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.8.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }


[tool.pdm]
distribution = false

[tool.pytest.ini_options]
pythonpath = [
    "src/"
]
[tool.ruff]
lint.extend-select = [
    "BLE",
    "C4",
    "I",
    "ICN",
    "ISC",
    "N",
    "NPY",
    "PGH",
    "PIE",
    "Q",
    "ASYNC",
    "RET",
    "RSE",
    "RUF",
    "SIM",
    "TID",
    "YTT"
]
exclude = [
    "alembic",
    "migrations"
]
lint.ignore = [
    "E402", "B905", "N818", "S101"
]
show-fixes = true